refactor(ingestion/dataplex): move build_lineage_parent to its only consumer
build_lineage_parent is a Data Lineage API concern used solely by
dataplex_lineage.py, not an identity primitive. Move it there (and its unit
test to test_dataplex_lineage.py) so dataplex_ids.py holds only mapper-facing
identity primitives (regexes, ContainerKey models, parse/instantiate helpers).
